Added basic hpack functionality needs tested
Added the ability to encode data into vectors of bytes representing
the indexed header fields and the literal header fields, also added
the ability to decode / encode integers, might remove the enum
representation for a series of functions and insted add a macro for
building the vectors of bytes. still need to write tests for a few of
the functions though.